Issue with long file names due to lossy type conversion in FtpServer::readChar() #75

+++ FtpServer.cpp
@@ -1902,7 +1902,7 @@ void FtpServer::abortTransfer()
 //     0 if empty line received
 //    length of cmdLine (positive) if no empty line received 
 
-int8_t FtpServer::readChar()
+int32_t FtpServer::readChar()
 {
   int8_t rc = -1;
